# src/docs/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Influencerium API",
        "description": "REST API for Influencerium - Influencer Data Management Platform. Manage influencers, campaigns, data models, and analytics with secure authentication and role-based access control.",
        "version": "1.0.0",
        "contact": {
            "name": "Influencerium Support",
            "email": "support@influencerium.com",
            "url": "https://influencerium.com/help"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "servers": [
        {
            "url": "http://localhost:3000/api/v1",
            "description": "Development server"
        },
        {
            "url": "https://api.influencerium.com/api/v1",
            "description": "Production server"
        }
    ],
...
    "paths": {
...
        "/models": {
            "get": {
                "tags": [
                    "Data Models"
                ],
                "summary": "List data models",
                "description": "Returns a paginated list of data models",
                "operationId": "listDataModels",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "search",
                        "schema": {
                            "type": "string"
                        },
                        "description": "Search by name or description"
                    },
                    {
                        "in": "query",
                        "name": "source",
                        "schema": {
                            "type": "string"
                        },
                        "description": "Filter by source"
                    },
                    {
                        "in": "query",
                        "name": "is_active",
                        "schema": {
                            "type": "boolean"
                        },
                        "description": "Filter by active status"
                    },
                    {
                        "in": "query",
                        "name": "page",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        },
                        "description": "Page number"
                    },
                    {
                        "in": "query",
                        "name": "limit",
                        "schema": {
                            "type": "integer",
                            "default": 20,
                            "maximum": 100
                        },
                        "description": "Items per page"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of data models",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DataModelListResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Data Models"
                ],
                "summary": "Create data model",
                "description": "Creates a new data model schema",
                "operationId": "createDataModel",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateDataModelRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Data model created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DataModel"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Model with this name already exists",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/models/{id}": {
            "get": {
                "tags": [
                    "Data Models"
                ],
                "summary": "Get data model",
                "description": "Returns a single data model by ID",
                "operationId": "getDataModel",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "description": "Data Model ID"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data model details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DataModel"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Data model not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Data Models"
                ],
                "summary": "Update data model",
                "description": "Updates an existing data model",
                "operationId": "updateDataModel",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "description": "Data Model ID"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateDataModelRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data model updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DataModel"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Data model not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Data Models"
                ],
                "summary": "Delete data model",
                "description": "Deletes a data model and all its data points",
                "operationId": "deleteDataModel",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "description": "Data Model ID"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data model deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Success"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Data model not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/models/{id}/data": {
            "get": {
                "tags": [
                    "Data Models"
                ],
                "summary": "List model data",
                "description": "Returns data points for a specific model",
                "operationId": "listModelData",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "description": "Data Model ID"
                    },
                    {
                        "in": "query",
                        "name": "page",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        },
                        "description": "Page number"
                    },
                    {
                        "in": "query",
                        "name": "limit",
                        "schema": {
                            "type": "integer",
                            "default": 50,
                            "maximum": 200
                        },
                        "description": "Items per page"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of data points",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ModelDataListResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Data Models"
                ],
                "summary": "Add data point",
                "description": "Adds a new data point to the model",
                "operationId": "addDataPoint",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "description": "Data Model ID"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AddDataPointRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Data point added successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DataPoint"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Data model not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "JWT authorization header. Example: `Authorization: Bearer <token>`"
            }
        },
        "schemas": {
            "Error": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "error": {
                        "type": "object",
                        "properties": {
                            "message": {
                                "type": "string",
                                "example": "An error occurred"
                            },
                            "code": {
                                "type": "string",
                                "example": "ERROR_CODE"
                            }
                        }
                    }
                }
            },
            "Success": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Operation successful"
                    }
                }
            },
...
            "DataModel": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string",
                        "example": "Instagram Influencer Profile"
                    },
                    "description": {
                        "type": "string"
                    },
                    "source": {
                        "type": "string",
                        "enum": [
                            "postedapp",
                            "instagram",
                            "tiktok",
                            "youtube",
                            "twitter",
                            "facebook",
                            "custom"
                        ]
                    },
                    "schema": {
                        "type": "object",
                        "description": "Model schema definition"
                    },
                    "settings": {
                        "type": "object"
                    },
                    "is_active": {
                        "type": "boolean"
                    },
                    "data_count": {
                        "type": "integer"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "CreateDataModelRequest": {
                "type": "object",
                "required": [
                    "name",
                    "source",
                    "schema"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "minLength": 3,
                        "maxLength": 100
                    },
                    "description": {
                        "type": "string",
                        "maxLength": 500
                    },
                    "source": {
                        "type": "string",
                        "enum": [
                            "postedapp",
                            "instagram",
                            "tiktok",
                            "youtube",
                            "twitter",
                            "facebook",
                            "custom"
                        ]
                    },
                    "schema": {
                        "type": "object",
                        "description": "Model schema definition"
                    },
                    "settings": {
                        "type": "object"
                    },
                    "is_active": {
                        "type": "boolean",
                        "default": true
                    }
                }
            },
            "UpdateDataModelRequest": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "minLength": 3,
                        "maxLength": 100
                    },
                    "description": {
                        "type": "string",
                        "maxLength": 500
                    },
                    "source": {
                        "type": "string",
                        "enum": [
                            "postedapp",
                            "instagram",
                            "tiktok",
                            "youtube",
                            "twitter",
                            "facebook",
                            "custom"
                        ]
                    },
                    "schema": {
                        "type": "object"
                    },
                    "settings": {
                        "type": "object"
                    },
                    "is_active": {
                        "type": "boolean"
                    }
                }
            },
            "DataModelListResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "data": {
                        "type": "object",
                        "properties": {
                            "models": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/DataModel"
                                }
                            },
                            "pagination": {
                                "$ref": "#/components/schemas/Pagination"
                            }
                        }
                    }
                }
            },
            "ModelDataListResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "data": {
                        "type": "object",
                        "properties": {
                            "model_name": {
                                "type": "string"
                            },
                            "data_points": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/DataPoint"
                                }
                            },
                            "pagination": {
                                "$ref": "#/components/schemas/Pagination"
                            }
                        }
                    }
                }
            },
            "DataPoint": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "model_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "data": {
                        "type": "object"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "AddDataPointRequest": {
                "type": "object",
                "required": [
                    "data"
                ],
                "properties": {
                    "data": {
                        "type": "object",
                        "description": "Data point object matching model schema"
                    }
                }
            },
...
            "Pagination": {
                "type": "object",
                "properties": {
                    "page": {
                        "type": "integer",
                        "example": 1
                    },
                    "limit": {
                        "type": "integer",
                        "example": 20
                    },
                    "total": {
                        "type": "integer",
                        "example": 100
                    },
                    "pages": {
                        "type": "integer",
                        "example": 5
                    }
                }
            }
        }
    }
}